Fix #278 Implement a requestid identifier to correlate queries
Let's use this to pass to idler so we can correlate the action between idler and
proxy.

We change the logging facility from Debug to Info for the Idle/Unidle but leave
it as Debug for State so we don't fill up too much the logs with noise.

Only testing was added for the new function. Writting tests for the whole
clients would need quite a bit of refactoring to use intefaces.