Fix #3390 Redirect to auth instead of showing 504
This patch fixes the 504 issue that happens while processing token_json
which is passed back from Auth. The root cause of 504 is that there
are calls to external (tenant, jenkins) services and was using the
default http-client which doesn't timeout. In case of network errors
these calls took more time  Openshift gateway's timeout (30 sec).
Thus causing the 504 Gateway error.

The solution for this included in the current patch is to use the
http-client from `utils` package which timeout after 15 seconds
generating an error. On error, the user is redirected to auth.
This however introduces the `ERR_TOO_MANY_REDIRECTS` but is less
of a problem than 504 as after showing the error, the browser does
retry the request.