Fix #315 Refactor auth to a separate package
This patch refactors Auth/Token related methods into its own package
to make it easier to write tests and to use the package. This refactor
adds an auth client which maintains the url to the auth service and
the public keys exposed by the service.

The SetDefaultClient and the DefaultClient functions makes it easier
to decouple auth from other packages. E.g. currently proxy
class/instance holds all the public keys (implementation details)
and passes that to utils funcs that require the public key(s).

With this change, public-keys will be owned by the auth.Client and
won't be exposed to callers.