Issue #321 Check if Session cookie is Valid
Currently, if session cookie is invalid, proxy does not check its
validity and passes the request to Jenkins considering that user is
logged in. Jenkins finds that cookie is not valid, but it can't expire
that cookie since it was setup by proxy. It will try to relogin user by
redirecting which will pass directly to jenkins again since session
cookie exists. And thus we are in a loop, leading to too many redirect
situation.

In proxy, check if cookie is valid. If it is not expire the cookie and
clear relevant cache. This will ask user to relogin without directly
passing the request to jenkins.