Handle errors instead of redirecting
When there is an errors instead of redirecting forever, just show it and exit.

This avoid the catch all issues "maximum redirect" and show a real error for
users to use instead.

There is no test for ui_requests but I have added a simple one that test a part
of this PR, there is much more to do to be able to do ui_requests properly.

This closes #330 and openshiftio/openshift.io#4180