Issue #321 #336 Check if Session cookie is Valid and Unit Tests (#335)
* Issue #321 Check if Session cookie is Valid

- Currently, if session cookie is invalid, proxy does not check its
validity and passes the request to Jenkins considering that user is
logged in. Jenkins finds that cookie is not valid, but it can't expire
that cookie since it was setup by proxy. It will try to relogin user by
redirecting which will pass directly to jenkins again since session
cookie exists. And thus we are in a loop, leading to too many redirect
situation.

Whenever jenkins given a Forbidden status, check if the session is
valid, if not expire cookie, delete the cache and redirect for login


* Issue #336 Tests for UI request and reverse proxy

- In order to make proxy package testable bundle dependency services
together.

We can have two logical services,
Jenkins: Contains methods that deals with a pool of Jenkins
Codebase: Contains methods that deals with getting information related
to the codebase

Use interfaces so that During test, we can create mock implementations
for this two interfaces, thus making the code testable

Apart from just making proxy testable, this change makes the code more
readable, better compartmentalizes it and uses Jenkins and Codebase
abstraction

- Avoid multiple write header situation by not writing header on
Forbidden

Added tests for
- jenkins http status vs reverse proxy status
- session validity being checked on recieving Forbidden from jenkins

- Use interface as an argument instead of struct for tenant and auth
- While expiring cookie `Set-Cookie` header was being written twice
(bug!), changed that
- handle error when default auth client is nil
- Added a mock auth service
- Added more tests for ui request, a) no token should redirect to auth
service, b) correct token, but jenkins is idled, should redirect to the
same url which was requested, set jenkins idled cookie, expire session
cookies if any

- Test if proxyCache is being filled
Added more tests, a) jenkins is running, but login fails because of
inappropriate token, b) jenkins is running and login is successful
(happy path)

- Added one more test
- proxy should expire the cookie if there is no cache entry with key
cookie value